@@ -12,50 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@ng-select/ng-select": "^3.7.1",
-    "@ngrx/effects": "^8.6.0",
-    "@ngrx/entity": "^8.6.0",
-    "@ngrx/store": "^8.6.0",
-    "@ngrx/store-devtools": "^8.6.0",
+    "@angular/animations": "~9.1.0",
+    "@angular/common": "~9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/core": "~9.1.0",
+    "@angular/forms": "~9.1.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
+    "@ng-select/ng-select": "^3.7.3",
+    "@ngrx/effects": "^9.0.0",
+    "@ngrx/entity": "^9.0.0",
+    "@ngrx/store": "^9.0.0",
+    "@ngrx/store-devtools": "^9.0.0",
     "bootstrap": "^4.4.1",
     "express": "^4.17.1",
-    "flag-icon-css": "^3.4.5",
-    "google-libphonenumber": "^3.2.6",
+    "flag-icon-css": "^3.4.6",
+    "google-libphonenumber": "^3.2.7",
     "jquery": "^3.4.1",
-    "ngx-bootstrap": "^5.2.0",
-    "rxjs": "~6.4.0",
+    "ngx-bootstrap": "^5.5.0",
+    "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.21",
-    "@angular/cli": "~8.3.21",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@ngrx/schematics": "^8.6.0",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
+    "@ngrx/schematics": "^9.0.0",
+    "@types/jasmine": "~3.5.9",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^8.9.5",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
+    "@types/node": "^13.9.1",
+    "codelyzer": "^5.2.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "protractor": "~5.4.3",
     "sw-precache-webpack-plugin": "^0.11.5",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "ts-node": "~8.6.2",
+    "tslint": "~6.1.0",
+    "typescript": "~3.8.3"
   }
 }
