@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute, NavigationEnd, Params, PRIMARY_OUTLET } from '@angular/router';
-import { filter } from 'rxjs/operators';
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Router, ActivatedRoute, NavigationEnd, Params, PRIMARY_OUTLET, NavigationStart, NavigationCancel } from '@angular/router';
+import { filter, takeWhile } from 'rxjs/operators';
+import { Location } from '@angular/common';
 
 interface BreadcrumbInterface {
   label: string;
@@ -13,29 +14,43 @@ interface BreadcrumbInterface {
   templateUrl: './breadcrumb.component.html',
   styleUrls: ['./breadcrumb.component.css']
 })
-export class BreadcrumbComponent implements OnInit {
+export class BreadcrumbComponent implements OnInit, OnDestroy {
 
   public breadcrumbs: BreadcrumbInterface[];
 
+  showSpinner: boolean;
+  componentIsActive: boolean;
 
   constructor(
     private activatedRoute: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private location: Location
   ) {
     this.breadcrumbs = [];
   }
 
   ngOnInit() {
+    this.componentIsActive = true;
     this.breadcrumbs = this.getBreadcrumbs(this.router.routerState.root);
-    const ROUTE_DATA_BREADCRUMB = 'breadcrumb';
+    // const ROUTE_DATA_BREADCRUMB = 'breadcrumb';
 
-    this.router.events.pipe(filter(event => event instanceof NavigationEnd)).subscribe(event => {
-      const root: ActivatedRoute = this.activatedRoute.root;
-      this.breadcrumbs = this.getBreadcrumbs(root);
-    });
+    this.router.events.pipe(filter(event => event instanceof NavigationEnd || event instanceof NavigationCancel))
+      .pipe(takeWhile(() => this.componentIsActive))
+      .subscribe(() => {
+        this.showSpinner = false;
+        const root: ActivatedRoute = this.activatedRoute.root;
+        this.breadcrumbs = this.getBreadcrumbs(root);
+      });
+    this.router.events.pipe(filter(event => event instanceof NavigationStart))
+      .pipe(takeWhile(() => this.componentIsActive))
+      .subscribe(() => {
+        this.showSpinner = true;
+      });
   }
 
-  private getBreadcrumbs(route: ActivatedRoute, url: string = '', breadcrumbs: BreadcrumbInterface[] = []): BreadcrumbInterface[] {
+  private getBreadcrumbs(
+    route: ActivatedRoute, url: string = '',
+    breadcrumbs: BreadcrumbInterface[] = []): BreadcrumbInterface[] {
     const ROUTE_DATA_BREADCRUMB = 'breadcrumb';
 
     const children: ActivatedRoute[] = route.children;
@@ -67,6 +82,16 @@ export class BreadcrumbComponent implements OnInit {
       }
       return this.getBreadcrumbs(child, url, breadcrumbs);
     }
+    return [];
+  }
+  backClicked() {
+    this.location.back();
+  }
+  goFullScreen() {
+    (document.querySelector('#main') as HTMLElement).requestFullscreen();
+  }
+  ngOnDestroy() {
+    this.componentIsActive = false;
   }
 
 }
