@@ -31,18 +31,19 @@ export class InputComponent implements OnInit, OnChanges, ControlValueAccessor {
   @Input() type: string;
   @Input() labelClass: string;
   @Input() inputClass: string;
-  @Input() ariaDescribedby: string;
-  fieldError: string;
+  @Input() step: number;
+  @Input() min: number;
+  fieldError: string | null;
   fieldType: string;
   disabled: boolean;
-  onChanges: ($value) => void;
+  onChanges: ($value: any) => void;
   onTouched: () => void;
   inputValue: any;
-  constructor(private appFormService: AppFormService) {}
+  constructor(private appFormService: AppFormService) { }
 
   ngOnInit() {
     this.fieldType = 'text';
-    if (['tel', 'phone', 'password', 'number', 'date'].includes(this.type)) {
+    if (['tel', 'phone', 'password', 'number', 'date', 'datetime-local'].includes(this.type)) {
       this.fieldType = this.type;
     }
     this.formControl = new FormControl();
@@ -52,6 +53,7 @@ export class InputComponent implements OnInit, OnChanges, ControlValueAccessor {
       const validationResult = this.formControl.validator(this.formControl);
       return (validationResult !== null && validationResult.required === true);
     }
+    return false;
   }
 
   setDisabledState?(isDisabled: boolean): void {
@@ -85,10 +87,11 @@ export class InputComponent implements OnInit, OnChanges, ControlValueAccessor {
   registerOnTouched(fn: any): void {
     this.onTouched = fn;
   }
-  get fieldClassIsValid() {
+  get fieldClassIsValid(): string | null {
     if (this.fieldError) {
-      return `is-invalid`;
+      return 'is-invalid';
     }
+    return null;
   }
   get fieldClass() {
     const formControlClass = 'form-control';
