@@ -34,7 +34,7 @@ describe('MenuSearchComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(MenuSearchComponent);
     component = fixture.componentInstance;
-    store = TestBed.get<Store<AppState>>(Store);
+    store = TestBed.inject<Store<AppState>>(Store);
 
     spyOn(store, 'dispatch').and.callThrough();
     fixture.detectChanges();
