@@ -36,7 +36,7 @@ describe('PasswordResetComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(PasswordResetComponent);
     component = fixture.componentInstance;
-    store = TestBed.get<Store<AppState>>(Store);
+    store = TestBed.inject<Store<AppState>>(Store);
 
     spyOn(store, 'dispatch').and.callThrough();
     fixture.detectChanges();
