@@ -33,7 +33,7 @@ describe('UserButtonComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(UserButtonComponent);
     component = fixture.componentInstance;
-    store = TestBed.get<Store<AppState>>(Store);
+    store = TestBed.inject<Store<AppState>>(Store);
 
     spyOn(store, 'dispatch').and.callThrough();
     fixture.detectChanges();
