@@ -8,15 +8,15 @@ describe('AuthenticationServiceService', () => {
   }));
 
   it('should be created', () => {
-    const service: AuthenticationService = TestBed.get(AuthenticationService);
+    const service: AuthenticationService = TestBed.inject(AuthenticationService);
     expect(service).toBeTruthy();
   });
   it('should have function "resetPassword', () => {
-    const service: AuthenticationService = TestBed.get(AuthenticationService);
+    const service: AuthenticationService = TestBed.inject(AuthenticationService);
     expect(service.resetPassword({email: 'some@valid.email'})).toBeTruthy();
   });
   it('should have function "login', () => {
-    const service: AuthenticationService = TestBed.get(AuthenticationService);
+    const service: AuthenticationService = TestBed.inject(AuthenticationService);
     expect(service.login({username: 'some@valid.email', password: 'someValidPassword', rememberMe: false})).toBeTruthy();
   });
 });
