@@ -2,8 +2,6 @@ export const environment = {
   production: true
 };
 
-export const PASSPORT_CLIENT = {
-  grantType: 'password',
-  clientSecret: 'lEo5ai7oDcSEPjz9sSsUqpvC5Oqlp8tMluGn3K31',
-  clientId: '4'
-};
+export const PASSPORT_CLIENT_GRANT_TYPE = 'password';
+export const PASSPORT_CLIENT_CLIENT_ID = '4';
+export const PASSPORT_CLIENT_CLIENT_SECRET = 'lEo5ai7oDcSEPjz9sSsUqpvC5Oqlp8tMluGn3K31';
