@@ -6,11 +6,9 @@ export const environment = {
   production: false
 };
 
-export const PASSPORT_CLIENT = {
-  grantType: 'password',
-  clientSecret: 'lEo5ai7oDcSEPjz9sSsUqpvC5Oqlp8tMluGn3K31',
-  clientId: '4'
-};
+export const PASSPORT_CLIENT_GRANT_TYPE = 'password'
+export const PASSPORT_CLIENT_CLIENT_ID = '4';
+export const PASSPORT_CLIENT_CLIENT_SECRET = 'lEo5ai7oDcSEPjz9sSsUqpvC5Oqlp8tMluGn3K31';
 
 /*
  * For easier debugging in development mode, you can import the following file
